sources/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.9.1
    # via django
attrs==25.3.0
    # via
    #   cattrs
    #   requests-cache
cattrs==25.1.1
    # via requests-cache
certifi==2025.7.14
    # via
    #   -r requirements.in
    #   opensearch-py
    #   requests
    #   sentry-sdk
charset-normalizer==3.4.2
    # via requests
django==5.2.4
    # via
    #   -r requirements.in
    #   django-health-check
    #   django-js-asset
    #   django-munigeo
    #   django-parler
    #   django-parler-rest
    #   djangorestframework
django-health-check==3.20.0
    # via -r requirements.in
django-js-asset==3.1.2
    # via django-mptt
django-mptt==0.17.0
    # via django-munigeo
django-munigeo==0.3.12
    # via -r requirements.in
django-parler==2.3
    # via
    #   django-munigeo
    #   django-parler-rest
django-parler-rest==2.2
    # via django-munigeo
djangorestframework==3.16.0
    # via django-parler-rest
events==0.5
    # via opensearch-py
idna==3.10
    # via
    #   requests
    #   url-normalize
opensearch-py==3.0.0
    # via -r requirements.in
platformdirs==4.3.8
    # via requests-cache
pyhumps==3.8.0
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via opensearch-py
pyyaml==6.0.2
    # via django-munigeo
requests==2.32.4
    # via
    #   -r requirements.in
    #   django-munigeo
    #   opensearch-py
    #   requests-cache
requests-cache==1.2.1
    # via django-munigeo
sentry-sdk==2.33.2
    # via -r requirements.in
six==1.17.0
    # via
    #   django-munigeo
    #   python-dateutil
sqlparse==0.5.3
    # via django
typing-extensions==4.14.1
    # via cattrs
url-normalize==2.2.1
    # via requests-cache
urllib3==2.5.0
    # via
    #   opensearch-py
    #   requests
    #   requests-cache
    #   sentry-sdk
